clean Server

+++ Content.Server/Mining/MiningSystem.cs
@@ -24,7 +24,7 @@ public sealed class MiningSystem : EntitySystem
     [Dependency] private readonly IPrototypeManager _proto = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
 
-    /// <inheritdoc/>
+    /// <inheritdoc />
     public override void Initialize()
     {
         base.Initialize();
@@ -44,7 +44,9 @@ public sealed class MiningSystem : EntitySystem
 
         var coords = Transform(uid).Coordinates;
         var modifier = component.Modifier; // Goobstation
-        var toSpawn = (int) _random.NextFloat((float) proto.MinOreYield * modifier, (float) (proto.MaxOreYield+1) * modifier); // Goobstation - Apply fortune
+        var toSpawn =
+            (int) _random.NextFloat(proto.MinOreYield * modifier,
+                (proto.MaxOreYield + 1) * modifier); // Goobstation - Apply fortune
         for (var i = 0; i < toSpawn; i++)
         {
             Spawn(proto.OreEntity, coords.Offset(_random.NextVector2(0.2f)));
@@ -53,7 +55,8 @@ public sealed class MiningSystem : EntitySystem
 
     private void OnMapInit(EntityUid uid, OreVeinComponent component, MapInitEvent args)
     {
-        if (component.CurrentOre != null || component.OreRarityPrototypeId == null || !_random.Prob(component.OreChance))
+        if (component.CurrentOre != null || component.OreRarityPrototypeId == null ||
+            !_random.Prob(component.OreChance))
             return;
 
         component.CurrentOre = _proto.Index<WeightedRandomOrePrototype>(component.OreRarityPrototypeId).Pick(_random);
